pyproject.toml
[project]
name = "arxiv-agent-crawler"
version = "0.1.0"
description = "AI-powered hybrid search engine for arXiv papers"
requires-python = ">=3.11"
dependencies = [
    # Orchestration
    "langgraph>=0.2",
    "langchain-core>=0.3",
    # Search & Retrieval
    "arxiv>=2.1",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.8",
    "sentence-transformers>=3.0",
    # Reranking
    "flashrank>=0.2",
    # LLM
    "litellm>=1.40",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    # Observability
    "langfuse>=2.40",
    # Resiliency
    "tenacity>=8.3",
    # Config
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "responses>=0.25",
    "ruff>=0.5",
    "black>=24.4",
]
desktop = [
    "PyQt6>=6.7",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
target-version = ["py311"]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
